backend/apply_migration.py
import os
from supabase import create_client, Client
from dotenv import load_dotenv

load_dotenv()

url: str = os.environ.get("SUPABASE_URL")
key: str = os.environ.get("SUPABASE_KEY")
supabase: Client = create_client(url, key)

def run_migration():
    with open("citizen_users.sql", "r") as f:
        sql = f.read()
    
    # Split by statement to execute one by one (Supabase client might not handle multiple statements in one call via rpc, 
    # but since we don't have direct SQL access, we can try to use a workaround or just print it for the user if this fails.
    # Actually, the python client doesn't have a direct 'query' method for raw SQL unless we have a stored procedure.
    # But wait, I can't create a stored procedure without SQL access.
    # I will try to use the 'rpc' if there is a generic sql runner, but unlikely.
    # ALTERNATIVE: I will ask the user to run it in their Supabase SQL Editor.
    # BUT, I can try to use the 'rest' interface to create the table? No.
    
    # Let's try to see if I can use the 'postgres' connection string if available? No.
    
    # Wait, I can use the 'run_command' to execute psql if installed? 
    # The user environment seems to be Windows.
    
    print("Please run the following SQL in your Supabase SQL Editor:")
    print(sql)

if __name__ == "__main__":
    run_migration()
