backend/requirements.txt
fastapi
fastapi
uvicorn
python-dotenv
supabase
httpx
google-generativeai
pydantic
google-adk
